+++ ParallelQSortOmenMP/ArrayFunc.h
@@ -2,7 +2,7 @@
 class ArrayFunc
 {
 public:
-    static bool PRKK(double *a, int n);
-    static void ShiftMas(double* _pData, int position, int shift, int c_size);
+    static bool PRKK(int *a, int n);
+    static void ShiftMas(int* _pData, int position, int shift, int c_size);
 };
